+++ src/PublicHolidays/Holidays/RelatedHolidays.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Swe\SpaceSDK\PublicHolidays\Holidays;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\AbstractApi;
+
+/**
+ * Class RelatedHolidays
+ *
+ * @package Swe\SpaceSDK\PublicHolidays\Holidays
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class RelatedHolidays extends AbstractApi
+{
+    /**
+     * Search related holidays in all public holiday calendars, during the selected period.
+     *
+     * @param array $request
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     */
+    public function getAllRelatedHolidays(array $request = [], array $response = []): array
+    {
+        $uri = 'public-holidays/holidays/related-holidays';
+
+        return $this->client->get($this->buildUrl($uri), $response, $request);
+    }
+}